requirements.txt
streamlit==1.56.0
PyMuPDF==1.27.2.2
pytesseract==0.3.13
Pillow==12.2.0
gTTS==2.5.4
mutagen==1.47.0
openai==2.30.0
numpy>=1.26.4,<3.0.0
requests>=2.32.0
langdetect==1.0.9
googletrans==4.0.2
httpx>=0.24.0
python-docx>=1.2.0
# cache-bust: 20260426-193622